requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.8.3
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cmweather==0.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dask==2025.9.1
deprecation==2.1.0
flatbuffers==25.9.23
fonttools==4.60.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h5netcdf==1.6.4
h5py==3.14.0
idna==3.10
keras==3.11.3
kiwisolver==1.4.9
lat_lon_parser==1.3.1
libclang==18.1.1
locket==1.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
netCDF4==1.7.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
partd==1.4.2
pillow==11.3.0
protobuf==5.29.5
Pygments==2.19.2
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scipy==1.16.2
setuptools==78.1.1
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
toolz==1.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wradlib==2.4.0
wrapt==1.17.3
xarray==2025.9.0
xmltodict==1.0.2
xradar==0.10.0
